Refactored FCM logic into a new FirebaseMessagingClient class
